# packages/api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "夸夸镜硬件 API",
    "version": "0.0.1",
    "description": "设备激活、配置、制品、OTA 与百炼模型接口。"
  },
  "servers": [
    {
      "url": "https://kuakua-api.cyrusdoyle.me",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/praise/stream": {
      "post": {
        "tags": ["百炼"],
        "summary": "图像夸夸 SSE",
        "parameters": [{ "$ref": "#/components/parameters/DeviceHeader" }, { "$ref": "#/components/parameters/K10ContentLength" }],
        "security": [{ "DeviceToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PraiseRequest" } } } },
        "responses": { "200": { "description": "text/event-stream，严格顺序：status → 全部 text → audio → complete；audio.url 指向 PCM 16-bit little-endian、单声道、16 kHz WAV。" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "411": { "$ref": "#/components/responses/ContentLengthRequired" }, "413": { "$ref": "#/components/responses/RequestTooLarge" }, "429": { "$ref": "#/components/responses/RateLimited" } }
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "tags": ["百炼"],
        "summary": "语音聊天 SSE",
        "parameters": [{ "$ref": "#/components/parameters/DeviceHeader" }, { "$ref": "#/components/parameters/K10ContentLength" }],
        "security": [{ "DeviceToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatRequest" } } } },
        "responses": { "200": { "description": "text/event-stream，严格顺序：status → asr_result → status → 全部 text → audio → complete；audio.url 指向 PCM 16-bit little-endian、单声道、16 kHz WAV。" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "411": { "$ref": "#/components/responses/ContentLengthRequired" }, "413": { "$ref": "#/components/responses/RequestTooLarge" }, "429": { "$ref": "#/components/responses/RateLimited" } }
      }
    },
    "/api/v1/tts": {
      "post": {
        "tags": ["百炼"],
        "summary": "文本转语音",
        "parameters": [{ "$ref": "#/components/parameters/DeviceHeader" }, { "$ref": "#/components/parameters/K10ContentLength" }],
        "security": [{ "DeviceToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TtsRequest" } } } },
        "responses": { "200": { "description": "临时 PCM 16-bit little-endian、单声道、16 kHz WAV 地址；设备以 Bearer Token 下载。" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "411": { "$ref": "#/components/responses/ContentLengthRequired" }, "413": { "$ref": "#/components/responses/RequestTooLarge" }, "429": { "$ref": "#/components/responses/RateLimited" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "DeviceToken": { "type": "http", "scheme": "bearer", "bearerFormat": "DeviceToken" }
    },
    "parameters": {
...
      "DeviceHeader": { "name": "X-Device-ID", "in": "header", "required": true, "schema": { "type": "string" } },
      "K10ContentLength": { "name": "Content-Length", "in": "header", "required": true, "description": "K10 模型接口的 JSON 请求体最大 1,500,000 字节。", "schema": { "type": "integer", "minimum": 1, "maximum": 1500000 } }
    },
    "responses": {
      "BadRequest": { "description": "请求不合法" },
      "Unauthorized": { "description": "DeviceToken 缺失、无效或设备归属不一致" },
      "RateLimited": { "description": "每设备模型调用超过 10 次/分钟" },
      "ContentLengthRequired": { "description": "K10 模型接口必须提供 Content-Length" },
      "RequestTooLarge": { "description": "K10 模型接口的 JSON 请求体超过 1,500,000 字节" },
...
    },
    "schemas": {
...
      "PraiseRequest": { "type": "object", "required": ["device_id", "image_base64"], "properties": { "device_id": { "type": "string" }, "image_base64": { "type": "string", "contentEncoding": "base64", "contentMediaType": "image/jpeg", "maxLength": 682668, "description": "无 Data URI 前缀、无换行的纯 JPEG Base64；解码后最大 500 KiB。" }, "timestamp": { "type": "integer", "format": "int64" } } },
      "ChatRequest": { "type": "object", "required": ["device_id", "audio_base64", "session_id"], "properties": { "device_id": { "type": "string" }, "audio_base64": { "type": "string", "contentEncoding": "base64", "contentMediaType": "audio/wav", "maxLength": 1398104, "description": "无 Data URI 前缀、无换行的完整 RIFF/WAV：PCM 16-bit little-endian、单声道、16 kHz；解码后最大 1 MiB。" }, "session_id": { "type": "string" }, "timestamp": { "type": "integer", "format": "int64" } } },
      "TtsRequest": { "type": "object", "required": ["device_id", "text"], "properties": { "device_id": { "type": "string" }, "text": { "type": "string", "minLength": 1, "maxLength": 100 } } }
    }
  }
}